refactor : params 정의
