+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/params/content/memo/MemoEditParameter.kt
@@ -1,6 +1,6 @@
-package com.whatever.caramel.core.domain.vo.memo
+package com.whatever.caramel.core.domain.params.content.memo
 
-import com.whatever.caramel.core.domain.vo.common.DateTimeInfo
+import com.whatever.caramel.core.domain.vo.content.schedule.DateTimeInfo
 import com.whatever.caramel.core.domain.vo.content.ContentAssignee
 
 data class MemoEditParameter(
