+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/params/content/schdule/ScheduleEditParameter.kt
@@ -1,6 +1,6 @@
-package com.whatever.caramel.core.domain.vo.calendar
+package com.whatever.caramel.core.domain.params.content.schdule
 
-import com.whatever.caramel.core.domain.vo.common.DateTimeInfo
+import com.whatever.caramel.core.domain.vo.content.schedule.DateTimeInfo
 import com.whatever.caramel.core.domain.vo.content.ContentAssignee
 
 data class ScheduleEditParameter(
